requirements.txt
Flask~=3.0.2
py-solc-x # geth compiler
retry==0.9.2
requests~=2.31.0
web3==6.15.1 # very important bug fixed in .1
